Moving BADGE_ICON_MAP
